syntastica

Module language_set

Defines the LanguageSet trait and some related types.

Also re-exports syntastica_highlight::HighlightConfiguration, tree_sitter::Language, and tft::FileType.

Structs

HighlightConfiguration

Contains the data needed to highlight code written in a particular language.

Language

An opaque object that defines how to parse a particular language. The code for each Language is generated by the Tree-sitter CLI.

Union

A combination of two arbitrary LanguageSets into one.

Enums

EitherLang

Either one of two language types.

FileType

A non-exhaustive list of text file types.

Traits

LanguageSet

Describes a type that is able to provide tree-sitter parsers and queries.

SupportedLanguage

A language included in a LanguageSet.