syntastica::renderer

Function resolve_styles

pub fn resolve_styles<'src>(
    highlights: &Highlights<'src>,
    theme: impl Borrow<ResolvedTheme>,
) -> ThemedHighlights<'src>

Convert Highlights to ThemedHighlights by applying styles from the given ResolvedTheme.

This may be useful if you do not want to render to a String and thus cannot use the Renderer trait for rendering, but still wish to only look up the theme styles once and then store the styled highlights.