Struct syntastica::renderer::HtmlRenderer

pub struct HtmlRenderer;

A Renderer implementor for HTML output. Use this type with the render function.

The resulting HTML is not wrapped in any parent tag and defines no font by itself. Each styled text region uses a span tag and defines its styling inline using the style attribute.

Example

use syntastica::{renderer::HtmlRenderer, theme};

// define a theme
let theme = theme! { "keyword": "#ff0000" }.resolve_links().unwrap();

// define highlights
let highlights = vec![vec![("<fn>", Some("keyword")), ("none", None)]];

// render to HTML
let output = syntastica::render(&highlights, &mut HtmlRenderer, theme);

assert_eq!(
    output,
    r#"<span style="color:rgb(255,0,0);">&lt;fn&gt;</span>none"#
);

Implementations

impl HtmlRenderer

pub fn new() -> Self

Create a new HtmlRenderer. Because HtmlRenderer is a unit struct, this is the same as just writing HtmlRenderer, but this function is provided for clarification nonetheless.