Enum syntastica::theme::ThemeValue

pub enum ThemeValue {
    Simple(String),
    Extended {
        color: Option<String>,
        underline: bool,
        strikethrough: bool,
        italic: bool,
        bold: bool,
        link: Option<String>,
    },
}

A value of a Theme containing style information and/or a link to another key in the Theme.

When using the serde feature, this type implements serde’s Serialize and Deserialize traits using the untagged enum representation.

Variants

Simple(String)

May either be a hexadecimal color literal, or a $ followed by the name of another theme key.

In the latter case, this value links to the ThemeValue which the Theme specifies for the provided theme key.

Extended

Fields

color: Option<String>

The color to use for this style, specified as a hexadecimal string.

Either this or link has to be set, or calls to Theme::resolve_links will fail.

underline: bool

Whether the text should be underlined. (default is false)

strikethrough: bool

Whether the text should be strikethrough. (default is false)

italic: bool

Whether the text should be italic. (default is false)

bold: bool

Whether the text should be bold. (default is false)

link: Option<String>

A link to the theme entry with the given key.

Either this or color has to be set, or calls to Theme::resolve_links will fail.

A color or link with additional style information.