Struct syntastica::theme::Theme

pub struct Theme(_);

A raw theme which may contain links to other items inside.

Internally, this type stores a map from Strings to ThemeValues. This map can be retrieved using Theme::into_inner. The map keys do not all have to be in THEME_KEYS; other custom keys can be used, for example to define a set of colors and reuse them with links everywhere else.

When using the serde feature, this type implements serde’s Serialize and Deserialize traits.

Instantiation

The easiest way to create a Theme is with the theme! macro. Alternatively, a Theme may be created from a BTreeMap<String, ThemeValue> using Theme::new.

Implementations

impl Theme

pub fn new(highlights: BTreeMap<String, ThemeValue, Global>) -> Theme

Create a new Theme from a map of theme keys to ThemeValues.

pub fn into_inner(self) -> BTreeMap<String, ThemeValue, Global>

Consume self and return the contained theme map.

May be used to merge multiple Themes together.

pub fn resolve_links(self) -> Result<ResolvedTheme, Error>

Try to resolve all links in this Theme and return a ResolvedTheme.

Errors

The function may return the following errors:

Error::InvalidHex if a color string was an invalid hexadecimal literal.

Error::InvalidLink if a link points to a non-existent key.