syntastica::ts_runtime::ffi

Function ts_wasm_store_load_language

pub unsafe extern "C" fn ts_wasm_store_load_language(
    arg1: *mut TSWasmStore,
    name: *const i8,
    wasm: *const i8,
    wasm_len: u32,
    error: *mut TSWasmError,
) -> *const TSLanguage

Create a language from a buffer of Wasm. The resulting language behaves like any other Tree-sitter language, except that in order to use it with a parser, that parser must have a Wasm store. Note that the language can be used with any Wasm store, it doesn’t need to be the same store that was used to originally load it.