syntastica

Type Alias ThemedHighlights

pub type ThemedHighlights<'src> = Vec<Vec<(&'src str, Option<Style>)>>;

Source code with theme-specific style information attached.

The type is a vector of lines of the source code, and each line is a vector of 2-tuples. The first element of each 2-tuple is a slice of the source text, and the second element is an optional style for that region of text.

Aliased Type

struct ThemedHighlights<'src> { /* private fields */ }