Crate syntastica::ts_runtime

Structs

IncludedRangesError

An error that occurred in Parser::set_included_ranges.

InputEdit

A summary of a change to a text document.

Language

An opaque object that defines how to parse a particular language. The code for each Language is generated by the Tree-sitter CLI.

LanguageError

An error that occurred when trying to assign an incompatible Language to a Parser.

LossyUtf8

Node

A single node within a syntax Tree.

Parser

A stateful object that this is used to produce a Tree based on some source code.

Point

A position in a multi-line text document, in terms of rows and columns.

Query

A set of patterns that match nodes in a syntax tree.

QueryCapture

A particular Node that has been captured with a particular name within a Query.

QueryCaptures

A sequence of QueryCaptures associated with a given QueryCursor.

QueryCursor

A stateful object for executing a Query on a syntax Tree.

QueryError

An error that occurred when trying to create a Query.

QueryMatch

A match of a Query to a particular set of Nodes.

QueryMatches

A sequence of QueryMatches associated with a given QueryCursor.

QueryPredicate

A key-value pair associated with a particular pattern in a Query.

QueryProperty

A key-value pair associated with a particular pattern in a Query.

Range

A range of positions in a multi-line text document, both in terms of bytes and of rows and columns.

Tree

A tree that represents the syntactic structure of a source code file.

TreeCursor

A stateful object for walking a syntax Tree efficiently.

Enums

CaptureQuantifier

A quantifier for captures

LogType

A type of log message.

QueryErrorKind

QueryPredicateArg

Constants

LANGUAGE_VERSION

The latest ABI version that is supported by the current version of the library.

MIN_COMPATIBLE_LANGUAGE_VERSION

The earliest ABI version that is supported by the current version of the library.

PARSER_HEADER

Traits

TextProvider

Functions

set_allocator⚠